requirements.txt
customtkinter==5.2.2
keras==3.0.5
matplotlib==3.8.3
numpy==1.26.4
opencv_python==4.9.0.80
Pillow==10.2.0
pytesseract==0.3.10
scikit_learn==1.4.1.post1
tensorflow==2.16.1
tensorflow_intel==2.16.1
